backend/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.12.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
